afpd: use fixed-width AFP text encoding hints
AFPName carries a four-byte, big-endian text encoding hint. Restore
v_kTextEncoding to uint32_t (which was mistakenly reverted in commit
d6224b7) and use the same fixed-width type in the charset registry,
removing architecture-dependent long storage and narrowing casts.

Expand kTextEncoding_t to cover the complete AFP 3.4 text encoding
identifier list. Document that defining an identifier does not imply
that Netatalk implements a converter for it, and make the composed
UTF-8 constant explicitly uint32_t.

Change set_name()'s final argument to bool because it selects legacy
versus UTF-8 name serialization; it never supplies the encoding value.
Update callers to pass true or false and continue copying the
volume's cached network-order hint for AFPName output.

Convert UTF-8 encoding hints with htonl() in the AFP test packet
helpers so their generated AFPNames are big-endian on little-endian
hosts.

+++ include/atalk/unicode.h
@@ -65,7 +65,7 @@ typedef enum {CH_UCS2 = 0, CH_UTF8 = 1, CH_MAC = 2, CH_UNIX = 3, CH_UTF8_MAC = 4
 
 struct charset_functions {
     const char *name;
-    const long kTextEncoding;
+    const uint32_t kTextEncoding;
     size_t (*pull)(void *, char **inbuf, size_t *inbytesleft,
                    char **outbuf, size_t *outbytesleft);
     size_t (*push)(void *, char **inbuf, size_t *inbytesleft,
